修复对比型信息图数据结构问题
- 更新Prompt，添加对比型模板数据结构说明和示例
- 在generate_service中添加_convert_to_comparison_structure方法
- 自动将LLM生成的扁平items数组转换为对比型所需的两层结构
- 智能分组策略：前半部分作为左侧，后半部分作为右侧
- 尝试从label中智能提取左右标题（方案A/B、iPhone/Android等）

解决了用户输入对比型文字后信息图无法正确显示的问题

+++ backend/app/models/work.py
@@ -2,7 +2,7 @@
 用户作品数据库模型
 """
 from datetime import datetime
-from sqlalchemy import Column, String, Text, BigInteger, DateTime, JSON, ForeignKey, Index
+from sqlalchemy import Column, String, Text, Integer, DateTime, JSON, ForeignKey, Index
 from app.models.base import Base
 
 
@@ -11,7 +11,7 @@ class UserWork(Base):
     __tablename__ = "user_works"
     
     # 字段定义
-    id = Column(BigInteger, primary_key=True, autoincrement=True, comment="作品ID")
+    id = Column(Integer, primary_key=True, autoincrement=True, comment="作品ID")
     user_id = Column(String(100), index=True, nullable=True, comment="用户标识")
     title = Column(String(200), nullable=True, comment="作品标题")
     template_id = Column(String(100), ForeignKey('templates.id'), nullable=False, comment="使用的模板ID")
